Авторизация
- Настроил Spring Security

+++ src/main/java/com/example/todo/service/TaskServiceImpl.java
@@ -1,12 +1,12 @@
 package com.example.todo.service;
 
+import com.example.todo.exception.NoMatchException;
 import com.example.todo.models.TaskDescription;
 import com.example.todo.models.TaskEntity;
-import com.example.todo.repository.CustomTaskRepository;
 import com.example.todo.repository.TaskRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -15,17 +15,23 @@ import java.util.NoSuchElementException;
 public class TaskServiceImpl implements TaskService {
 
     private final TaskRepository repository;
+    private final AccountService userService;
 
     public void toggleTask(Long id) {
+
         TaskEntity task = repository.findById(id).orElseThrow(() ->
                 new NoSuchElementException("Задачи с таким id: " + id + "не существует"));
+        checkTaskOwner(task);
         task.setCompleted(!task.getCompleted());
         repository.save(task);
     }
 
+
     public void editTask(Long id, TaskDescription taskDescription) {
+
         TaskEntity task = repository.findById(id).orElseThrow(() ->
                 new NoSuchElementException("Задачи с таким id: " + id + "не существует"));
+        checkTaskOwner(task);
         task.setDescription(taskDescription.getTaskDescription());
         task.setCompleted(false);
         repository.save(task);
@@ -33,18 +39,30 @@ public class TaskServiceImpl implements TaskService {
 
 
     public List<TaskEntity> getTasks(Boolean all, String substring) {
-        return repository.getTasks(all, substring);
+
+        return repository.getTasks(all, substring,
+                userService.getAccountByUsername(SecurityContextHolder.getContext().getAuthentication().getName()).getId());
     }
 
     @Override
-    public void addTask(TaskDescription description) {
-        repository.save(new TaskEntity(description, false));
+    public void addTask(TaskDescription description, String username) {
+
+        repository.save(new TaskEntity(description, false, userService.getAccountByUsername(username)));
     }
 
     @Override
     public void deleteTask(Long id) {
+
         TaskEntity task = repository.findById(id).orElseThrow(() ->
                 new NoSuchElementException("Задачи с таким id: " + id + "не существует"));
+        checkTaskOwner(task);
         repository.delete(task);
+    }
+
+    private void checkTaskOwner(TaskEntity task) {
+
+        if (!task.getAccount().getUsername().equals(SecurityContextHolder.getContext().getAuthentication().getName())) {
+            throw new NoMatchException("Задача из параллельной вселенной, технологии не освоены");
+        }
     }
 }
